fix Makefile clean rule, fix include path in c files

+++ src/str/ft_strnequ.c
@@ -6,15 +6,15 @@
 /*   By: rseelaen <rseelaen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/27 21:19:36 by rseelaen          #+#    #+#             */
-/*   Updated: 2023/05/02 13:56:49 by rseelaen         ###   ########.fr       */
+/*   Updated: 2023/05/19 14:50:27 by rseelaen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "../../libft.h"
 
 int	ft_strnequ(char const *s1, char const *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while ((s1[i] || s2[i]) && i < n)
